Agregar nuevo hospedaje - Afiliado
Faltan detalles.

+++ Proyecto_Avanzada/Entities/HospedajeEnt.cs
@@ -14,7 +14,7 @@ namespace Proyecto_Avanzada.Entities
     {
         public long ConsecutivoHospedaje { get; set; }
         public string Nombre { get; set; }
-        public byte CodProvincia { get; set; }
+        public short CodProvincia { get; set; }
         public string NombreProvincia { get; set; }
 
         public float Precio { get; set; }
